Check /var/dump existence
Before removing contents of /var/dump the shotgun now checks that the
directory exists.

Closes-Bug: #1574695

+++ shotgun/manager.py
@@ -83,5 +83,6 @@ class Manager(object):
             msg = "Clearing target failed. function: {}, path: {}, excinfo: {}"
             logger.error(msg.format(function, path, excinfo))
 
-        shutil.rmtree(os.path.dirname(self.conf.target),
-                      onerror=on_rmtree_error)
+        if os.path.exists(self.conf.target):
+            shutil.rmtree(os.path.dirname(self.conf.target),
+                        onerror=on_rmtree_error)
